+++ python2/raygun4py/raygunmsgs.py
@@ -1,7 +1,6 @@
-import traceback
-import inspect
-import os
 import sys
+import os
+import inspect
 
 try:
     import multiprocessing
@@ -28,7 +27,7 @@ class RaygunMessageBuilder:
         self.raygunMessage.details['machineName'] = name
         return self
 
-    def set_environment_details(self):
+    def set_environment_details(self, extra_environment_data):
         self.raygunMessage.details['environment'] = {
             "processorCount": (
                 multiprocessing.cpu_count() if USE_MULTIPROCESSING else "n/a"
@@ -36,8 +35,15 @@ class RaygunMessageBuilder:
             "architecture": platform.architecture()[0],
             "cpu": platform.processor(),
             "oSVersion": "%s %s" % (platform.system(), platform.release()),
-            "environmentVariables": os.environ.data
+            "environmentVariables": os.environ.data,
+            "runtimeLocation": sys.executable,
+            "runtimeVersion": 'Python ' + sys.version
         }
+
+        if extra_environment_data is not None:
+            merged = extra_environment_data.copy()
+            merged.update(self.raygunMessage.details['environment'])
+            self.raygunMessage.details['environment'] = merged
 
         return self
 
@@ -48,14 +54,14 @@ class RaygunMessageBuilder:
     def set_client_details(self):
         self.raygunMessage.details['client'] = {
             "name": "raygun4py",
-            "version": "2.2.0",
+            "version": "3.1.0",
             "clientUrl": "https://github.com/MindscapeHQ/raygun4py"
         }
         return self
 
-    def set_customdata(self, userCustomData):
-        if type(userCustomData) is dict:
-            self.raygunMessage.details['userCustomData'] = userCustomData
+    def set_customdata(self, user_custom_data):
+        if type(user_custom_data) is dict:
+            self.raygunMessage.details['userCustomData'] = user_custom_data
         return self
 
     def set_tags(self, tags):
@@ -72,7 +78,7 @@ class RaygunMessageBuilder:
               "queryString": request['queryString'],
               "form": request['form'],
               "headers": request['headers'],
-              "rawData": request['rawData'],
+              "rawData": request['rawData']
             }
 
             if 'ipAddress' in request: