+++ python2/tests/test_functional.py
@@ -37,6 +37,16 @@ class TestRaygun4PyFunctional(unittest.TestCase):
 
             self.assertEqual(httpResult[0], 202)
 
+    def test_send_with_tags(self):
+        client = raygunprovider.RaygunSender(self.apiKey)
+
+        try:
+            raise Exception("Raygun4py manual sending test - tags")
+        except:
+            httpResult = client.send_exception(exc_info=sys.exc_info(), tags=["I am a tag"])
+
+            self.assertEqual(httpResult[0], 202)
+
     def log_send(self, logger):
         try:
             raise StandardError("Raygun4py Logging Test")