+++ python3/raygun4py/middleware/flask.py
@@ -1,10 +1,6 @@
-from __future__ import absolute_import
-
 import logging
-
 import flask
 from flask.signals import got_request_exception
-
 from raygun4py import raygunprovider
 
 log = logging.getLogger(__name__)