+++ python3/raygun4py/middleware/wsgi.py
@@ -1,7 +1,5 @@
 import logging
-
 from raygun4py import raygunprovider
-
 
 log = logging.getLogger(__name__)